Address review comments
- Move tests to shared common test locations and add a group to
make it clear whether we're running on the web or not.
- Remove dynamic in favor of Object? in module metadata.
- Add constants for field names in module metadata.
- Add documentation for json files and parameters.
- Change named of restart and reload scripts.
