+++ packages/flutter_tools/test/web.shard/test_data/hot_restart_web_test_common.dart
@@ -72,11 +72,13 @@ Future<void> _testProject(
   late Directory tempDir;
   late FlutterRunTestDriver flutter;
 
-  final String testName = 'Hot restart (index.html: $name)';
   final List<String> additionalCommandArgs =
       useDDCLibraryBundleFormat
           ? <String>['--extra-front-end-options=--dartdevc-canary,--dartdevc-module-format=ddc']
           : <String>[];
+  final String testName =
+      'Hot restart (index.html: $name)'
+      '${additionalCommandArgs.isEmpty ? '' : ' with args: $additionalCommandArgs'}';
 
   setUp(() async {
     tempDir = createResolvedTempDirectorySync('hot_restart_test.');
